+++ backend/corpora/common/corpora_orm.py
@@ -6,7 +6,6 @@ import sys
 from sqlalchemy import (
     Boolean,
     Column,
-    create_engine,
     DateTime,
     Enum,
     Float,
@@ -17,12 +16,11 @@ from sqlalchemy import (
 )
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.ext.declarative import declarative_base, DeclarativeMeta
-from sqlalchemy.orm import relationship, sessionmaker
+from sqlalchemy.orm import relationship
 
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from .corpora_config import CorporaDbConfig
 from .utils.exceptions import CorporaException
 
 
@@ -83,18 +81,6 @@ class AuditMixin(object):
 
 
 Base = declarative_base(cls=TransformingBase)
-
-
-class DBSessionMaker:
-    engine = create_engine(CorporaDbConfig().database_uri, connect_args={"connect_timeout": 5})
-    _session_make = None
-
-    def __init__(self):
-        if not self._session_make:
-            self._session_make = sessionmaker(bind=self.engine)
-
-    def session(self, **kwargs):
-        return self._session_make(**kwargs)
 
 
 class CollectionVisibility(enum.Enum):
