+++ tests/unit/backend/chalice/api_server/test_v1_dataset.py
@@ -5,7 +5,7 @@ import unittest
 from furl import furl
 
 from backend.corpora.common.corpora_orm import UploadStatus
-from backend.corpora.common.utils.db_utils import processing_status_updater
+from backend.corpora.common.utils.db_session import processing_status_updater
 from tests.unit.backend.chalice.api_server.base_api_test import BaseAuthAPITest
 from tests.unit.backend.chalice.api_server.mock_auth import get_auth_token
 from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
