+++ tests/unit/backend/fixtures/data_portal_test_case.py
@@ -1,6 +1,6 @@
 import unittest
 
-from backend.corpora.common.corpora_orm import DBSessionMaker
+from backend.corpora.common.utils.db_session import DBSessionMaker
 from tests.unit.backend.fixtures.generate_data_mixin import GenerateDataMixin
 from tests.unit.backend.fixtures.test_db import TestDatabaseManager
