+++ tests/unit/backend/fixtures/test_db.py
@@ -13,8 +13,8 @@ from backend.corpora.common.corpora_orm import (
     ValidationStatus,
     ConversionStatus,
     ProcessingStatus,
-    DBSessionMaker,
 )
+from backend.corpora.common.utils.db_session import DBSessionMaker
 from backend.scripts.create_db import create_db
 from tests.unit.backend.fixtures import config
